bluebell 帖子投票功能完善

+++ logic/vote.go
@@ -36,6 +36,6 @@ func VoteForPost(userID int64, p *models.ParamVoteData) error {
 	zap.L().Debug("VoteForPost",
 		zap.Int64("userID", userID),
 		zap.String("postID", p.PostID),
-		zap.Float64("direction", p.Direction))
-	return redis.VoteForPost(strconv.Itoa(int(userID)), p.PostID, p.Direction)
+		zap.Int8("direction", p.Direction))
+	return redis.VoteForPost(strconv.Itoa(int(userID)), p.PostID, float64(p.Direction))
 }
